core/api/viewsets.py
from rest_framework import viewsets
from core.api import serializers
from core import models

class PacienteViewSet(viewsets.ModelViewSet):
    serializer_class = serializers.PacienteSerializer
    queryset = models.Paciente.objects.all()