requirements.txt
asgiref==3.4.1
boto3==1.19.2
botocore==1.22.2
Django==3.2.8
djangorestframework==3.12.4
jmespath==0.10.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.2.0
gunicorn==20.1.0
psycopg2==2.8.6
django-environ==0.8.1    
django-storages==1.12.2
django-cors-headers==3.10.0